Add a new stddev_scaling_factor setting.
Users can now influence the width of the Gaussian distribution
function, using a new stddev_scaling_factor setting. This allows
the user to adjust how strongly the algoithm will prefer items at
the top of the list over items at the bottom of the list. A larger
number will give items at the top of the list a stronger chance of
being chosen.

The setting is optional, defaulting to 3.0.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>