Define a default for model so users don't have to.
Prior to this commit, users had to define the model setting, and
it had to be set to "gaussian". This was silly, especially since
rpick does not yet support any other model.

This commit defines "gaussian" as a default, so users don't have
to enter it in their configs. It still can only be set to default,
for now.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>