Don't prompt user about a choice they've declined.
It was annoying if rpick suggested a choice the user has already
declined, so this commit adds a list of declined choices to the
Engine. We don't want the API user to have to deal with this, so
this was a good opportunity to make the Engine struct members
private and provide a new() function to replace it.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>