Make the tests pass on 32-bit arches.
The RNG picks different values on 32-bit arches than it does on
64-bit arches. This commit adjusts our tests to seed the RNGs
with different values on 32-bit arches so that they pick the same
results as the 64-bit arches do for tests.

This is not an awesome way to solve the problem, but it is quick.
A better solution would be to write a fake RNG that implements the
expected trait and use that in the tests instead. However, this
will work for now.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>