Refactor and fix the consent code
Prior to this commit, if users had a config that had a category
with items in it that had 0 chance of being chosen and they said
no to every choice that was picked, rpick would enter an infinite
loop. This condition was likely to happen with the lottery model
since the most recently chosen item is assigned 0 tickets.

Furthermore, the library engine was storing state for items that
the user said no to without associating the items with a category,
which would prevent reuse of the engine for more than one rpick.

This commit refactors how the consent is requested and handled to
solve both of these issues.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>