Add a test RNG that is arch-independent
Prior to this commit, we were using SmallRng with hardcoded seeds
to get deterministic test results. This led to problems when
running the tests on 32-bit architectures, as they would cause the
RNG to pick different values in the tests and cause the assertions
to fail.

This commit adds a new FakeRng for testing that always returns the
values 0, 1, 2…, no matter whether the host is a 32-bit or 64-bit
system. This way the test results on both architectures end up
picking the same values, which makes testing a bit easier to
maintain going forward.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>