Skip a doctest on 32-bit architectures
32-bit architectures end up with different PRNG results than
64-bit systems, and this results in test failure on them. It's
annoying to try to find a seed value on 32-bit systems that ends
up with the same results on 64-bit systems, so we simply disable
the doctest on 32-bit systems.

In the tests we used a fake PRNG to ensure the same results
between the two classes of architectures, but that would be
cumbersome to use in the doctests, because the purpose there is to
show users how to use the API and that code would be distracting.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>