Introduce a new UI trait and alter API to use it
This is a backwards breaking change in the crate.

The Engine now receives only an instance of a struct that
implements the new UI trait. The UI trait is friendlier towards API
users than streams of text.

Instantiating an Engine no longer accepts an Rng. There is a new
set_rng() method on the Engine if you wish to provide your own
Rng.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>