Refactor to drop the workspace
I found a way to use the include!() macro instead of splitting the CLI
struct out into its own crate. This has a few oddities of its own, like
rust-analyzer not helping when you edit command.rs and use statements
being a bit odd, but it does avoid publishing a crate that only exists
to facilitate building a man page.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>